Revert "refactoring path to files"
This reverts commit 7288c5cfe2b692215c01fec198d2e6ded7b1018b.
