+++ app.js
@@ -3,8 +3,7 @@ const logger = require('morgan')
 const cors = require('cors')
 
 const authRouter = require('./routes/api/auth')
-const userRouter = require('./routes/api/users')
-const contactsRouter = require('./routes/api/contacts')
+const contactsRouter = require('./routes/api/index')
 
 const app = express()
 
@@ -15,7 +14,6 @@ app.use(cors())
 app.use(express.json())
 
 app.use('/api/auth', authRouter)
-app.use('/api/users', userRouter)
 app.use('/api/contacts', contactsRouter)
 
 app.use((req, res) => {
