+++ src/controllers/auth/index.js
@@ -1,5 +1,5 @@
-const register = require('./register')
-const login = require('./login')
+const { register } = require('./register')
+const { login } = require('./login')
 
 module.exports = {
   register,
