+++ src/controllers/auth/login.js
@@ -22,4 +22,4 @@ const login = async (req, res) => {
   }  
 }
 
-module.exports = login
+module.exports = { login }
