+++ src/controllers/auth/register.js
@@ -31,4 +31,4 @@ const register = async (req, res) => {
     }
 }
 
-module.exports = register
+module.exports = { register }
