+++ src/controllers/contacts/deleteContact.js
@@ -13,4 +13,6 @@ const deleteContact = async (req, res) => {
     })    
 }
 
-module.exports = deleteContact
+module.exports = {
+  deleteContact
+}
