+++ src/controllers/contacts/getAll.js
@@ -5,4 +5,6 @@ const getList = async (req, res) => {
   return res.json({ status: 'success', code: 200, contacts: data })
 }
 
-module.exports = getList
+module.exports = {
+  getList
+}
