+++ src/controllers/contacts/getContact.js
@@ -8,4 +8,6 @@ const getContact = async (req, res, next) => {
     .catch(_ => res.status(404).json({ message: `Contact with id=${contactId} not found`, code: 404, status: 'falure' }))
 }
 
-module.exports = getContact
+module.exports = {
+  getContact
+}
