--- src/controllers/contacts/index.js
@@ -1,15 +1,0 @@
-const addContact = require('./addContact')
-const deleteContact = require('./deleteContact')
-const getAll = require('./getAll')
-const getContact = require('./getContact')
-const updateContact = require('./updateContact')
-const updateStatus = require('./updateStatus')
-
-module.exports = {
-  addContact,
-  deleteContact,
-  getAll,
-  getContact,
-  updateContact,
-  updateStatus
-}
