+++ src/controllers/contacts/updateContact.js
@@ -18,4 +18,6 @@ const updateContact = async (req, res) => {
     })
     .catch(err => res.status(400).json({ message: err.message, code: 400, status: 'falure' }))}
 
-module.exports = updateContact
+module.exports = {
+  updateContact
+}
