--- src/controllers/users/getCurrent.js
@@ -1,7 +1,0 @@
-// const { User } = require('../../../models')
-
-const getCurrent = async(req, res) => {
-  console.log(req.User)  
-}
-
-module.exports = getCurrent
