--- src/controllers/users/index.js
@@ -1,3 +1,0 @@
-const getCurrent = require('./getCurrent')
-
-module.exports = { getCurrent }
