--- src/middlewares/auth.js
@@ -1,7 +1,0 @@
-// const { User } = require('../../models')
-
-const auth = async(req, res) => {
-
-}
-
-module.exports = auth
