--- src/middlewares/index.js
@@ -1,4 +1,0 @@
-const validation = require('./validationMiddleware')
-const auth = require('./auth')
-
-module.exports = { auth, validation }
